pyproject.toml
[project]
name = "bantay-paws-app"
version = "0.1.0"
description = "Non-profit website with FastAPI and file-based content"
readme = "README.md"
authors = [
    { name = "Br0k3nDesign", email = "adam.j.moran@gmail.com" }
]
requires-python = ">=3.12"

dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "jinja2",
  "python-multipart",
  "pyyaml",
  "markdown",
  "python-dotenv",
  "starlette",
  "itsdangerous>=2.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
]

[project.scripts]
dev = "-- uvicorn app.main:app --reload"
